api/requirements.txt
Flask
psycopg2-binary
pika
